fixup consul-template-template-templates documentation
Fixing readme uses and ordering of
consul_template_template_templates. This was rather
confusing for me to see so many "template" words in a row, and the
readme for its usage should be fixed.

It makes the most sense to have this reordered straight under
consul_template_template_files as both are tasks for creating
consul-template templates.